pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "deep-agent-coder"
version = "0.1.0"
description = "Multi-agent coding system built on LangChain Deep Agents"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "you@example.com"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "deepagents>=0.2.0",
    "langchain-anthropic>=1.0.0",
    "langgraph-checkpoint-postgres>=3.0.0",
    "psycopg[binary]>=3.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[project.scripts]
deep-agent-coder = "deep_agent_coder.cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/deep-agent-coder"
Documentation = "https://github.com/yourusername/deep-agent-coder#readme"
Repository = "https://github.com/yourusername/deep-agent-coder"

[tool.setuptools.packages.find]
where = ["."]
include = ["deep_agent_coder*"]

[tool.pytest.ini_options]
markers = [
    "integration: marks tests that require ANTHROPIC_API_KEY and make real API calls",
]
testpaths = ["tests"]
